refactor: share CLI shortcut labels
