+++ src/test-kernel/cli/StatusBar.vitest.mts
@@ -7,7 +7,7 @@ import {
   statusBarSegmentText,
   statusBarStreamTarget,
 } from '@cli/chat/tui/panes/StatusBar';
-import { defaultShortcutModifierLabel } from '@cli/chat/tui/shortcutLabels';
+import { defaultShortcutModifierLabel } from '@cli/runtime/shortcutLabels';
 import { shortCliApiMode } from '@cli/runtime/apiAccessMode';
 import { NO_BYPASS, type StreamSlice } from '@cli/chat/tui/state/cliState';
 import { STREAM_STATUS } from '@shared/schemas';
